requirements.txt
dataclasses==0.6
Django==2.1.7
django-stubs==0.9.0
djangorestframework==3.9.2
djangorestframework-simplejwt==4.1.0
djangorestframework-stubs==0.3.0
future==0.17.1
mypy==0.670
mypy-extensions==0.4.1
PyJWT==1.7.1
pytz==2018.9
typed-ast==1.3.1
typing-extensions==3.7.2
